small refactoring
added api.js
changed default city
